+++ src/Business/Features/Lessons/Models/Lessons/GetLessonLiteModel.cs
@@ -2,6 +2,6 @@
 
 public sealed class GetLessonLiteModel : IResponseModel
 {
-    public byte Id { get; set; }
+    public short Id { get; set; }
     public string? Name { get; set; }
 }
