+++ src/Business/Features/Lessons/Models/Lessons/UpdateLessonModel.cs
@@ -2,7 +2,7 @@
 
 public sealed class UpdateLessonModel : IResponseModel
 {
-    public byte Id { get; set; }
+    public short Id { get; set; }
     public string? Name { get; set; }
-    public byte SortNo { get; set; }
+    public short SortNo { get; set; }
 }
