+++ src/Business/Features/Packages/Models/Packages/UpdatePackageModel.cs
@@ -4,7 +4,7 @@ public sealed class UpdatePackageModel : IResponseModel
 {
     public short Id { get; set; }
     public string? Name { get; set; }
-    public byte SortNo { get; set; }
+    public short SortNo { get; set; }
     public bool IsWebVisible { get; set; }
     public double TaxRatio { get; set; }
     public double Amount { get; set; }
