+++ src/Business/Features/Questions/Models/Questions/AddQuestionModel.cs
@@ -2,7 +2,7 @@
 
 public class AddQuestionModel : IRequestModel
 {
-    public byte LessonId { get; set; }
+    public short LessonId { get; set; }
     public string? QuestionPictureBase64 { get; set; }
     public string? QuestionPictureFileName { get; set; }
 }
