+++ src/Business/Features/Questions/Models/Quizzes/AddQuizModel.cs
@@ -2,7 +2,7 @@
 
 public class AddQuizModel : IRequestModel
 {
-    public byte LessonId { get; set; }
+    public short LessonId { get; set; }
     public long UserId { get; set; }
     public List<string> QuestionList { get; set; } = [];
     public List<bool> VisualList { get; set; } = [];
