+++ src/Business/Features/Schools/Models/ClassRooms/AddClassRoomModel.cs
@@ -4,5 +4,5 @@ public class AddClassRoomModel : IRequestModel
 {
     public short No { get; set; }
     public string? Branch { get; set; }
-    public byte PackageId { get; set; }
+    public short PackageId { get; set; }
 }
